# docs/swagger.yaml
info:
  contact: {}
  license: {}
paths:
  /api/auth/AddAuth:
    post:
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 姓名
        in: query
        name: name
        required: true
        type: string
      - description: 组：crm/calculate
        in: query
        name: group
        required: true
        type: string
      - description: 电话
        in: query
        name: telephone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
        "500":
          description: '{"code":500,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增个人用户
  /api/exchange/addMyExchange:
    post:
      parameters:
      - description: 申请对象
        in: query
        name: respondent
        required: true
        type: string
      - description: 申请时间
        in: query
        name: request_time
        required: true
        type: string
      - description: 被申请时间
        in: query
        name: requested_time
        required: true
        type: string
      - description: 换班类型，1，晚班，2,周末白班，3，crm工作日特殊班，4，周末全天班
        in: query
        name: exchange_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 新增本人换班申请表信息
...
  /api/exchange/deleteMyExchange{id}:
    post:
      parameters:
      - description: Id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 删除本人的未审批换班请求表
  /api/exchange/examineExchange:
    post:
      parameters:
      - description: Id
        in: query
        name: id
        required: true
        type: integer
      - description: 回复，状态 0为默认、1为拒绝、2为同意
        in: query
        name: response
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 回复换班申请表
...
  /api/exchange/getMyExamine:
    post:
      parameters:
      - description: 状态 0：未审批 1：已审批
        in: query
        name: state
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 查看本人回复的换班申请表信息(未审批/已审批)
  /api/exchange/myExchange:
    post:
      parameters:
      - description: 状态 0：未审批 1：已审批
        in: query
        name: state
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 查看本人的换班请求表(未审批/已审批)
...
  /auth:
    post:
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 通过用户名密码获取token
  /register:
    post:
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 名字
        in: query
        name: name
        required: true
        type: string
      - description: 组：crm/calculate
        in: query
        name: group
        required: true
        type: string
      - description: 电话
        in: query
        name: telephone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 注册个人用户
swagger: "2.0"